+++ swf-plugin-collab/src/main/java/com/venky/swf/plugins/collab/db/model/user/User.java
@@ -7,8 +7,8 @@ import com.venky.swf.db.annotations.column.ui.HIDDEN;
 public interface User extends com.venky.swf.plugins.security.db.model.User{
 	@Deprecated
 	@HIDDEN
-	public Integer getCompanyId(); 
-	public void setCompanyId(Integer companyId);
+	public Long getCompanyId();
+	public void setCompanyId(Long companyId);
 	
 	public List<UserCompany> getUserCompanies();
 }
