+++ swf-plugin-collab/src/main/java/com/venky/swf/plugins/collab/extensions/participation/UserFacilityParticipantExtension.java
@@ -17,10 +17,10 @@ public class UserFacilityParticipantExtension extends ParticipantExtension<UserF
 	}
 
 	@Override
-	protected List<Integer> getAllowedFieldValues(User user, UserFacility model, String fieldName) {
-		List<Integer> ret = null;
+	protected List<Long> getAllowedFieldValues(User user, UserFacility model, String fieldName) {
+		List<Long> ret = null;
 		if (fieldName.equalsIgnoreCase("FACILITY_ID")){
-			ret = new ArrayList<Integer>();
+			ret = new ArrayList<Long>();
 			if (model.getFacilityId() > 0 ){
 				if (model.getFacility().isAccessibleBy(user, Facility.class)){
 					if (model.getUserId() == 0 || model.getFacility().isAccessibleBy(model.getUser(), Facility.class)){
@@ -30,7 +30,7 @@ public class UserFacilityParticipantExtension extends ParticipantExtension<UserF
 			}else {
 				List<Facility> facilites = DataSecurityFilter.getRecordsAccessible(Facility.class,user);
  				if (model.getUserId() > 0 ){
- 					ret = new ArrayList<Integer>();
+ 					ret = new ArrayList<Long>();
 					for (Facility f : facilites){
 						if (f.isAccessibleBy(model.getUser(), Facility.class)){
 							ret.add(f.getId());
@@ -42,7 +42,7 @@ public class UserFacilityParticipantExtension extends ParticipantExtension<UserF
 			}
 		}else if (fieldName.equalsIgnoreCase("USER_ID")){
 			if (model.getUserId() > 0){
-				ret = new ArrayList<Integer>();
+				ret = new ArrayList<Long>();
 				if (model.getUser().isAccessibleBy(user, User.class)){
 					if (model.getFacilityId() == 0 || 
 							(model.getFacility().isAccessibleBy(user, Facility.class) && 
@@ -51,7 +51,7 @@ public class UserFacilityParticipantExtension extends ParticipantExtension<UserF
 					}
 				}
 			}else {
-				ret = new SequenceSet<Integer>();
+				ret = new SequenceSet<Long>();
 				if (model.getFacilityId() > 0 && !model.getFacility().isAccessibleBy(user)){
 					return ret; 
 				}
@@ -59,11 +59,11 @@ public class UserFacilityParticipantExtension extends ParticipantExtension<UserF
 				
 				Facility f = model.getFacility();
 				if (f != null ){ 
-					SequenceSet<Integer> subscribedUserIds = new SequenceSet<>();
+					SequenceSet<Long> subscribedUserIds = new SequenceSet<>();
 					for (FacilityUser fu : f.getFacilityUsers()){
 						subscribedUserIds.add(fu.getUserId());
 					}
-					SequenceSet<Integer> subscribableUserIds = new SequenceSet<>();
+					SequenceSet<Long> subscribableUserIds = new SequenceSet<>();
 					f.getCompany().getCompanyUsers().forEach((cu)->subscribableUserIds.add(cu.getUserId()));
 					subscribableUserIds.removeAll(subscribedUserIds);
 					ret.addAll(subscribableUserIds);
